requirements.txt
Pillow==8.1.2
ImageHash==4.2.0
fire==0.4.0
